requirements.txt
# Core Gmail and parsing libraries
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
mail-parser>=3.15.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html2text>=2020.1.16

# Duplicate detection and similarity
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
hdbscan>=0.8.0

# Utilities and patterns
backoff>=2.2.0
python-dateutil>=2.8.0

# Optional but recommended
pandas>=2.0.0
extruct>=0.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
